Fix detection of coredumps during testing
The test scripts were incorrectly detecting non-asterisk coredumps
and failing the tests.  We've added a custom ast_coredumper.sh
to help fix that and updated the test scripts to call it instead
of the version shipped with asterisk.  When testing this patch,
we also discovered that runAsteriskUnitTests.sh wasn't actually
generating coredumps when it should because of the way sudo works.
Since this script will always be run by root anyway, we've removed
the need for sudo.